Chert's leg detects the ground

+++ mv_from_locomo.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-mv ../TravelersLocomotionPack/TravelersLocomotionPack/assets/assetbundles/pickupchert* PickUpChert/assets/assetbundles/.
+if [ -e ../TravelersLocomotionPack/TravelersLocomotionPack/assets/assetbundles/pickupchert ]; then
+    echo "getting assetbundle of PickUpChert..."
+    mv ../TravelersLocomotionPack/TravelersLocomotionPack/assets/assetbundles/pickupchert* PickUpChert/assets/assetbundles/.
+fi
+echo "copying ILocomotionAPI.cs..."
+cp ../TravelersLocomotionPack/TravelersLocomotionPack/ILocomotionAPI.cs PickUpChert/.
